Use feature test macros properly in igs_resource_msg_from_err_unix.
Feature test macros should be used to inform the C library what
functions a program _expects_ to be available (and which should be made
available to it), not as checks for what functions are available to be
used. Therefore, it is necessary to set _POSIX_C_SOURCE to 200809L,
which indicates we wish to use functions defined in POSIX.2008, one of
which is `int strerror_r()`. This way, all libc (including glibc, as
long as _GNU_SOURCE is undefined) implementations will make the correct
version of the function available, and the fallback to the `char
*strerror_r()` version is no longer necessary.